+++ content/gallery.php
@@ -19,9 +19,9 @@
 				<time <?php hybrid_attr( 'entry-published' ); ?>><?php echo get_the_date(); ?></time>
 				<span class="post-by"><?php echo esc_html_x( 'by', 'post author', 'themelia' ) ?></span>
 				<span <?php hybrid_attr( 'entry-author' ); ?>><?php the_author_posts_link(); ?></span>
-				<?php if ( comments_open() && 0 < $number ) : '<span class="sep">' . _ex( ' | ', 'By Line separator', 'themelia' ) . '</span>'; endif ?>
-				<?php comments_popup_link( ( '' ), '<span>' . number_format_i18n( 1 ) . '</span>' . __( ' Comment', 'themelia' ), '<span>%</span>' .  __( ' Comments', 'themelia' ), 'comments-link' . $css_class, '' ); ?>
-				<?php edit_post_link(); ?>
+				<?php if ( comments_open() && 0 < $number ) : echo '<span class="sep">' . _x( ' | ', 'By Line separator', 'themelia' ) .  '</span>'; endif ?>
+				<?php comments_popup_link( '', '<span>1</span> comment', '<span>%</span> comments', 'comments-link', '' ); ?>
+				<?php themelia_edit_link(); ?>
 			</div><!-- .entry-byline -->
 
 			<h1 <?php hybrid_attr( 'entry-title' ); ?>><?php single_post_title(); ?></h1>
@@ -49,9 +49,9 @@
 				<time <?php hybrid_attr( 'entry-published' ); ?>><?php echo get_the_date(); ?></time>
 				<span class="post-by"><?php echo esc_html_x( 'by', 'post author', 'themelia' ) ?></span>
 				<span <?php hybrid_attr( 'entry-author' ); ?>><?php the_author_posts_link(); ?></span>
-				<?php if ( comments_open() && 0 < $number ) : '<span class="sep">' . _ex( ' | ', 'By Line separator', 'themelia' ) . '</span>'; endif ?>
-				<?php comments_popup_link( ( '' ), '<span>' . number_format_i18n( 1 ) . '</span>' . __( ' Comment', 'themelia' ), '<span>%</span>' .  __( ' Comments', 'themelia' ), 'comments-link' . $css_class, '' ); ?>
-				<?php edit_post_link(); ?>
+				<?php if ( comments_open() && 0 < $number ) : echo '<span class="sep">' . _x( ' | ', 'By Line separator', 'themelia' ) .  '</span>'; endif ?>
+				<?php comments_popup_link( '', '<span>1</span> comment', '<span>%</span> comments', 'comments-link', '' ); ?>
+				<?php themelia_edit_link(); ?>
 			</div><!-- .entry-byline -->
 
 			<?php the_title( '<h2 ' . hybrid_get_attr( 'entry-title' ) . '><a href="' . esc_url( get_permalink() ) . '" rel="bookmark" itemprop="url">', '</a></h2>' ); ?>