+++ content/link.php
@@ -21,9 +21,9 @@
 				<time <?php hybrid_attr( 'entry-published' ); ?>><?php echo get_the_date(); ?></time>
 				<span class="post-by"><?php echo esc_html_x( 'by', 'post author', 'themelia' ) ?></span>
 				<span <?php hybrid_attr( 'entry-author' ); ?>><?php the_author_posts_link(); ?></span>
-				<?php if ( comments_open() && 0 < $number  ) : '<span class="sep">' . _ex( ' | ', 'By Line separator', 'themelia' ) . '</span>'; endif ?>
-				<?php comments_popup_link( ( '' ), '<span>' . number_format_i18n( 1 ) . '</span>' . __( ' Comment', 'themelia' ), '<span>%</span>' .  __( ' Comments', 'themelia' ), 'comments-link' . $css_class, '' ); ?>
-				<?php edit_post_link(); ?>
+				<?php if ( comments_open() && 0 < $number ) : echo '<span class="sep">' . _x( ' | ', 'By Line separator', 'themelia' ) .  '</span>'; endif ?>
+				<?php comments_popup_link( '', '<span>1</span> comment', '<span>%</span> comments', 'comments-link', '' ); ?>
+				<?php themelia_edit_link(); ?>
 			</div><!-- .entry-byline -->
 
 		</header><!-- .entry-header -->
@@ -51,9 +51,9 @@
 			<time <?php hybrid_attr( 'entry-published' ); ?>><?php echo get_the_date(); ?></time>
 			<span class="post-by"><?php echo esc_html_x( 'by', 'post author', 'themelia' ) ?></span>
 			<span <?php hybrid_attr( 'entry-author' ); ?>><?php the_author_posts_link(); ?></span>
-			<?php if ( comments_open() && 0 < $number  ) : '<span class="sep">' . _ex( ' | ', 'By Line separator', 'themelia' ) . '</span>'; endif ?>
-			<?php comments_popup_link( ( '' ), '<span>' . number_format_i18n( 1 ) . '</span>' . __( ' Comment', 'themelia' ), '<span>%</span>' .  __( ' Comments', 'themelia' ), 'comments-link' . $css_class, '' ); ?>
-			<?php edit_post_link(); ?>
+			<?php if ( comments_open() && 0 < $number ) : echo '<span class="sep">' . _x( ' | ', 'By Line separator', 'themelia' ) .  '</span>'; endif ?>
+			<?php comments_popup_link( '', '<span>1</span> comment', '<span>%</span> comments', 'comments-link', '' ); ?>
+			<?php themelia_edit_link(); ?>
 		</div><!-- .entry-byline -->
 
 		<footer class="entry-footer small">